use plugin to register imod/relion
