--- cylindra/ext/__init__.py
@@ -1,5 +1,0 @@
-from cylindra.ext._utils import CommandExecutionError, CommandNotFound
-from cylindra.ext.imod import IMOD
-from cylindra.ext.relion import RELION
-
-__all__ = ["CommandExecutionError", "CommandNotFound", "IMOD", "RELION"]
