--- cylindra/ext/imod/__init__.py
@@ -1,4 +1,0 @@
-from .cmd import IMODCommand, read_mod, save_angles, save_mod
-from .menu import IMOD
-
-__all__ = ["read_mod", "save_mod", "save_angles", "IMOD", "IMODCommand"]
