--- cylindra/ext/relion/__init__.py
@@ -1,3 +1,0 @@
-from cylindra.ext.relion.menu import RELION
-
-__all__ = ["RELION"]
