+++ docs/api/builtins/relion.md
@@ -1,0 +1,8 @@
+# cylindra_builtins.relion
+
+This built-in plugin submodule provides functions to work with RELION file formats.
+
+::: cylindra_builtins.relion
+    options:
+        show_signature_annotations: false
+        heading_level: 4
