--- docs/api/ext.md
@@ -1,17 +1,0 @@
-# cylindra.ext
-
-This submodule includes functions for external softwares.
-
-## IMOD
-
-::: cylindra.ext.IMOD
-    options:
-        show_signature_annotations: false
-        heading_level: 4
-
-## RELION
-
-::: cylindra.ext.RELION
-    options:
-        show_signature_annotations: false
-        heading_level: 4
