+++ docs/extern/index.md
@@ -1,7 +1,8 @@
 # Working with Other Softwares
 
 To maximize the power of computational tools, `cylindra` provides several functions to
-import and export data from other softwares.
+import and export data from other softwares. These functions are implemented as built-in
+plugins using the `cylindra` [plugin system](../plugins.md).
 
 Currently supported softwares are:
